fix: also widen non-logger resource-count accumulators to long
The Sum-based overflow fixed in the loggers has sibling instances in the
other resource-count aggregators, which accumulate per-VC counts (each an
int) across a whole run:

- VerifyCommand / VerificationStatistics.TotalResourcesUsed
  (--performance-stats 'Total resources used'): an Interlocked.Add into an
  int field. Unlike Enumerable.Sum it does not throw on overflow -- it
  silently wraps to a wrong, possibly negative, total. Widen the field to
  long (the Interlocked.Add then binds the long overload and the int
  element widens) and widen the Round helper accordingly. MaxVcResourcesUsed
  stays int: it is a per-VC Math.Max, never a sum.
- MeasureComplexityCommand 'total consumed resources': an int '+=' with the
  same silent-wrap behaviour. Accumulate in long.

The IDE gutter roll-up was already safe (the tree node's ResourceCount field
is already long, and the PR widened the feeding parameter to long).

Note: a single implementation's count is still capped upstream, where Boogie
sums per-split counts into an int (SplitAndVerifyWorker.totalResourceCount);
that is outside this repository.

+++ Source/DafnyDriver/CliCompilation.cs
@@ -389,6 +389,6 @@ record VerificationStatistics {
   public int OutOfResourceCount;
   public int OutOfMemoryCount;
   public int SolverExceptionCount;
-  public int TotalResourcesUsed;
+  public long TotalResourcesUsed;
   public int MaxVcResourcesUsed;
 }
